refactor and split into different files
